+++ db/migrate/20190116110744_create_series.rb
@@ -5,7 +5,7 @@ class CreateSeries < ActiveRecord::Migration[5.2]
   def change
     create_table :series do |t|
       t.string :name
-      t.integer :budget
+      t.decimal :budget
       t.timestamps
     end
   end
